backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
web3==6.4.0
python-dotenv==1.0.0
